Lumbago_Music/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lumbago-music"
version = "1.0.0"
description = "Lumbago Music AI — DJ Library Manager"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Lumbago Dev" }]
dependencies = [
    "PyQt6>=6.6",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "librosa>=0.10",
    "numpy>=1.26",
    "pyloudnorm>=0.1",
    "pydub>=0.25",
    "mutagen>=1.47",
    "pyacoustid>=1.3",
    "openai>=1.30",
    "anthropic>=0.28",
    "google-generativeai>=0.7",
    "musicbrainzngs>=0.7",
    "discogs-client>=2.4",
    "pylast>=5.3",
    "spotipy>=2.23",
    "rapidfuzz>=3.9",
    "send2trash>=1.8",
    "watchdog>=4.0",
    "python-decouple>=3.8",
    "reportlab>=4.2",
    "Jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-qt>=4.4",
    "pytest-asyncio>=0.23",
    "pytest-benchmark>=4.0",
    "factory-boy>=3.3",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[project.scripts]
lumbago = "lumbago_app.ui.app:main"
lumbago-launcher = "launcher:main"

[tool.hatch.build.targets.wheel]
packages = ["lumbago_app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "PIE", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
qt_api = "pyqt6"
